feat: adding ssm construct
